requirements.txt
numpy==1.24.3 
pandas==2.0.3 
pytesseract==0.3.10 
pdf2image==1.16.3 
Pillow==10.0.0 
PyMuPDF==1.23.5 
openpyxl==3.1.2 
pyyaml==6.0.1 
schedule==1.2.0
opencv-python
PyPDF2==3.0.1
reportlab==4.0.4
pywin32==306; platform_system=="Windows"
watchdog==3.0.0


# Utilities
python-dateutil==2.8.2

